projects/svd_image_modes/requirements.txt
numpy
scipy
pillow
matplotlib
pandas
pyarrow
pyyaml
scikit-image
